Add support for local agent execution
Fixes #13

This commit adds a new execution mode called "--local".
This mode does not use Docker for isolation, and instead
executes the agent directly on the host machine.

In order to do this, the changes add support for checking for and
installing all of the software that the agent depends on.

Additionally, it changes the syntax of certain commands to make
them more cross-platform compatible.

It also removes the udp-punch program as the Wireguard
persistent keep-alive setting replaces the need for it.

**Tested**
- Verified the `--local` flag records and runs a test successfully.
- Verified without the flag that the Docker approach still works.
- Connected to a running agent and killed the `connect` script, verified it exits.
- Verified the agent exits when `monitor` dies while running locally.
- Verified Ctrl+C properly kills all sub-processes.
- Verified if I remove a dependency, the check script reports it.
- Verified that the installation script succeeds and removes extra files.

+++ run-agent.sh
@@ -3,8 +3,14 @@
 PublicPort=10009
 
 function usage {
-  echo "usage: $0 -k <reflect_api_key> [-p <public_port] [-n]"
+  echo "usage: $0 [--local] -k <reflect_api_key> [-p <public_port>]"
   echo -e "\tRuns the Reflect Agent and connects to the specified Reflect account"
+  echo
+  echo -e "\t--local"
+  echo -e "\t\tRuns the agent without Docker isolation on the local machine."
+  echo -e "\t\tThis requires installing several utility program dependencies."
+  echo -e "\t\tSee the 'local/check-dependency.sh' and 'local/install...' scripts."
+  echo -e "\t\tNOTE: this mode requires 'sudo' since it modifies network interfaces."
   echo
   echo -e "\t-k reflect_api_key"
   echo -e "\t\tThe API key for the Reflect account"
@@ -14,6 +20,11 @@ function usage {
   echo
   exit 1
 }
+
+if [ "$1" == "--local" ]; then
+  IsLocal="true"
+  shift
+fi
 
 while getopts ":k:p:" option; do
     case "${option}" in
@@ -34,11 +45,26 @@ if [ -z "${ReflectApiKey}" ]; then
     usage
 fi
 
-echo "Reflect Agent running on port $PublicPort"
+if [ "$IsLocal" == "true" ]; then
+  Directory=$(dirname -- $0)
+  $Directory/local/run-local.sh $ReflectApiKey $PublicPort
 
-docker run --rm --cap-add net_admin -d \
-  --name agent \
-  -e ReflectApiKey=$ReflectApiKey \
-  -e PublicPort=$PublicPort \
-  -p $PublicPort:$PublicPort/udp \
-  agent
+  LocalAgentExit=$?
+  if [ "$LocalAgentExit" != "0" ]; then
+    echo "Reflect Agent exited unexpectedly: $LocalAgentExit"
+  else
+    echo "Reflect Agent exited successfully"
+  fi
+
+else
+  echo "Reflect Agent running on port $PublicPort"
+
+  docker run --rm --cap-add net_admin -d \
+    --name agent \
+    -e ReflectApiKey=$ReflectApiKey \
+    -e PublicPort=$PublicPort \
+    -p $PublicPort:$PublicPort/udp \
+    agent
+
+  # The agent daemon is stopped using 'stop-agent.sh'.
+fi
